Дописал интеграционный тест на загрузку файла, пока есть проблема с ним

+++ backend/FileService/FileService.Core/FilesStorage/IChunkSizeCalculator.cs
@@ -5,5 +5,5 @@ namespace FileService.Core.FilesStorage;
 
 public interface IChunkSizeCalculator
 {
-    Result<(long ChunkSize, int TotalChunks), Error> CalculateChunkSize(long fileSize);
+    Result<(int ChunkSize, int TotalChunks), Error> CalculateChunkSize(long fileSize);
 }
